added keyboard input for cells when not in noteMode. noteMode is toggled with a right mouse click, and cells can be selected with left click
