+++ src/gui/backend/Settings.java
@@ -346,7 +346,7 @@ public class Settings {
         cellGUIStartMode = false;
         defaultOpenState = 0;
         theme = new Theme(new File(appDirectory + "default.theme"));
-        autoFillNotes = false;
+        autoFillNotes = true;
 
         updateSettingsFile();
     }
